Fixes to most warnings generated with SIERRA flags.

+++ packages/zoltan/src/params/key_params.c
@@ -454,10 +454,10 @@ void Zoltan_Print_Configuration(char *indent)
   if (indent == NULL){
     indent = "";
   }
-  printf("\n%sZOLTAN_ID_TYPE: %s (%zd bytes)\n",
-    indent, zoltan_id_datatype_name, sizeof(ZOLTAN_ID_TYPE));
-  printf("%sZOLTAN_GNO_TYPE: %s, (%zd bytes)\n",
-    indent, zoltan_gno_datatype_name, sizeof(ZOLTAN_GNO_TYPE));
+  printf("\n%sZOLTAN_ID_TYPE: %s (%lu bytes)\n",
+    indent, zoltan_id_datatype_name, (unsigned long)(sizeof(ZOLTAN_ID_TYPE)));
+  printf("%sZOLTAN_GNO_TYPE: %s, (%lu bytes)\n",
+    indent, zoltan_gno_datatype_name, (unsigned long)(sizeof(ZOLTAN_GNO_TYPE)));
   printf("%sMPI_Datatype for ZOLTAN_ID_TYPE: %s\n", indent,
     zoltan_mpi_id_datatype_name);
   printf("%sMPI_Datatype for ZOLTAN_GNO_TYPE: %s\n", indent,
